requirements.txt
yt-dlp
pywin32; platform_system == "Windows"
Pillow 